ref(rust-consumer): Move clickhouse_client out of rust_arroyo
There's no point having in the clickhouse_client living in the arroyo package
since it's only by the ClickHouse strategy in Snuba which is not part of Arroyo.
