+++ rust_snuba/rust_arroyo/Cargo.toml
@@ -20,6 +20,5 @@ cadence = "0.29.0"
 lazy_static = "1.4.0"
 parking_lot = "0.10.0"
 rand="0.8.5"
-reqwest = "0.11.11"
 serde_json = "1.0.81"
 serde = {version = "1.0.137", features = ["derive"] }
