+++ rust_snuba/rust_arroyo/src/utils/mod.rs
@@ -1,3 +1,2 @@
-pub mod clickhouse_client;
 pub mod clock;
 pub mod metrics;
