+++ rust_snuba/src/strategies/clickhouse.rs
@@ -1,11 +1,11 @@
 use std::time::Duration;
 use rust_arroyo::processing::strategies::{CommitRequest, MessageRejected, ProcessingStrategy};
 use rust_arroyo::types::{Message};
-use rust_arroyo::utils::clickhouse_client::ClickhouseClient;
 use crate::types::BytesInsertBatch;
 use crate::config::ClickhouseConfig;
 use tokio::task::JoinHandle;
-
+use reqwest::header::{HeaderMap, HeaderValue, ACCEPT_ENCODING, CONNECTION};
+use reqwest::{Client, Error, Response};
 
 pub struct ClickhouseWriterStep {
     next_step: Box<dyn ProcessingStrategy<BytesInsertBatch>>,
@@ -99,5 +99,61 @@ impl ProcessingStrategy<BytesInsertBatch> for ClickhouseWriterStep {
 
     fn join(&mut self, timeout: Option<Duration>) -> Option<CommitRequest> {
         self.next_step.join(timeout)
+    }
+}
+
+#[derive(Clone)]
+pub struct ClickhouseClient {
+    client: Client,
+    url: String,
+    headers: HeaderMap<HeaderValue>,
+    table: String,
+}
+impl ClickhouseClient {
+    pub fn new(hostname: &str, http_port: u16, table: &str, database: &str) -> ClickhouseClient {
+        let mut client = ClickhouseClient {
+            client: Client::new(),
+            url: format!("http://{}:{}", hostname, http_port),
+            headers: HeaderMap::new(),
+            table: table.to_string(),
+        };
+
+        client
+            .headers
+            .insert(CONNECTION, HeaderValue::from_static("keep-alive"));
+        client
+            .headers
+            .insert(ACCEPT_ENCODING, HeaderValue::from_static("gzip,deflate"));
+        client
+            .headers
+            .insert("X-ClickHouse-Database", HeaderValue::from_str(database).unwrap());
+        client
+    }
+
+    pub async fn send(&self, body: Vec<u8>) -> Result<Response, Error> {
+        self.client
+            .post(self.url.clone())
+            .headers(self.headers.clone())
+            .body(body)
+            .query(&[(
+                "query",
+                format!("INSERT INTO {} FORMAT JSONEachRow", self.table),
+            )])
+            .send()
+            .await
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[tokio::test]
+    async fn it_works() -> Result<(), reqwest::Error> {
+        let client: ClickhouseClient = ClickhouseClient::new("localhost", 8123, "querylog_local", "default");
+
+        println!("{}", "running test");
+        let res = client.send(b"[]".to_vec()).await;
+        println!("Response status {}", res.unwrap().status());
+        Ok(())
     }
 }
